Create group formation with parameters
An institution decides to allow students to create student groups for each subject where students with similar marks can help each other. But it draws up a set of constraints for creating student groups:

A group should be associated with a particular subject.
The difference between the scores of any two students in the group should be at most mark_limit.
It follows that multiple groups can be formed for a given subject.